pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "daylily-cognito"
dynamic = ["version"]
description = "Shared Cognito authentication library for FastAPI + Jinja2 web apps"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "Daylily Informatics", email = "daylily@daylilyinformatics.com"}
]
keywords = ["aws", "cognito", "authentication", "fastapi", "jwt", "oauth2"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",

    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: Session",
    "Typing :: Typed",
]
dependencies = [
    "boto3>=1.26.0",
    "cli-core-yo==1.1.0",
    "fastapi>=0.104.0",
    "itsdangerous>=2.1.0",
    "PyYAML>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
auth = [
    "python-jose[cryptography]>=3.3.0",
]
dev = [
    "bandit[toml]>=1.8.0",
    "httpx>=0.24.0",
    "pre-commit>=3.8.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
daycog = "daylily_cognito.cli:main"

[project.urls]
Homepage = "https://github.com/Daylily-Informatics/daylily-cognito"
Repository = "https://github.com/Daylily-Informatics/daylily-cognito"

[tool.setuptools.packages.find]
where = ["."]
include = ["daylily_cognito*"]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "E402",  # module level import not at top (needed for conditional imports in tests)
]

[tool.ruff.format]
quote-style = "double"

[tool.bandit]
targets = ["daylily_cognito"]
exclude_dirs = ["tests", "dist", "build"]
skips = ["B101", "B105", "B404", "B603", "B607"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.setuptools_scm]
# Derives version from git tags (e.g., 0.1.4 from tag "0.1.4")
# If commits exist after tag, version becomes "0.1.4.dev1+g<hash>"
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
